Remove modal and replace with external links
