+++ src/DashboardUI/src/components/SiteNavbar.tsx
@@ -11,9 +11,7 @@ import { useState } from 'react';
 interface SiteNavbarProps {
   currentTab: HomePageTab;
   onTabClick: (tab: HomePageTab) => void;
-  showAboutModal: (show: boolean) => void;
   showContactModal(show: boolean): void;
-  showTermsModal(show: boolean): void;
 }
 
 function SiteNavbar(props: SiteNavbarProps) {
@@ -73,11 +71,11 @@ function SiteNavbar(props: SiteNavbarProps) {
         </Offcanvas.Header>
         <Offcanvas.Body>
           <Nav defaultActiveKey="/" className="flex-column gap(10px)">
-            <Nav.Link onClick={() => props.showAboutModal(true)}>About</Nav.Link>
-            <Nav.Link onClick={() => { }}>Water Rights Data</Nav.Link>
-            <Nav.Link onClick={() => { }}>Aggregate Water Use Data</Nav.Link>
+            <Nav.Link target="_blank" href="https://westernstateswater.org/wade/about ">About</Nav.Link>
+            <Nav.Link target="_blank" href="https://westernstateswater.org/wade/water-rights-data">Water Rights Data</Nav.Link>
+            <Nav.Link target="_blank" href="https://westernstateswater.org/wade/Aggregate-Water-Data">Aggregate Water Use Data</Nav.Link>
             <Nav.Link onClick={() => props.showContactModal(true)}>Contact Us</Nav.Link>
-            <Nav.Link onClick={() => props.showTermsModal(true)}>Terms and Conditions</Nav.Link>
+            <Nav.Link target="_blank" href="https://westernstateswater.org/wade/terms-of-service">Terms and Conditions</Nav.Link>
           </Nav>
         </Offcanvas.Body>
       </Offcanvas>
