+++ src/DashboardUI/src/pages/HomePage.tsx
@@ -1,11 +1,9 @@
 import mapboxgl from 'mapbox-gl';
 import { useEffect, useState } from 'react';
-import AboutModal from '../components/AboutModal';
 import ContactModal from '../components/ContactModal';
 import SidePanel from '../components/SidePanel';
 import SiteFooter from '../components/SiteFooter';
 import SiteNavbar from '../components/SiteNavbar';
-import TermsModal from '../components/TermsModal';
 
 import '../styles/home-page.scss';
 
@@ -18,10 +16,7 @@ export enum HomePageTab {
 function HomePage() {
 
   const [currentTab, setCurrentTab] = useState(HomePageTab.WaterRights);
-
-  const [shouldAboutShowModal, setShowAboutModal] = useState(false);
   const [showContactModal, setShowContactModal] = useState(false);
-  const [showTermsModal, setShowTermsModal] = useState(false);
 
   useEffect(() => {
     mapboxgl.accessToken = process.env.REACT_APP_MAPBOX_ACCESSTOKEN || "";
@@ -37,21 +32,13 @@ function HomePage() {
     setCurrentTab(tab);
   }
 
-  const shouldShowAboutModal = (show: boolean) => {
-    setShowAboutModal(show);
-  }
-
   const shouldShowContactModal = (show: boolean) => {
     setShowContactModal(show);
   }
 
-  const shouldShowTermsModal = (show: boolean) => {
-    setShowTermsModal(show);
-  }
-
   return (
     <div className="home-page d-flex flex-column">
-      <SiteNavbar onTabClick={handleTabClick} currentTab={currentTab} showAboutModal={shouldShowAboutModal} showContactModal={shouldShowContactModal} showTermsModal={shouldShowTermsModal} />
+      <SiteNavbar onTabClick={handleTabClick} currentTab={currentTab} showContactModal={shouldShowContactModal} />
       <div className="d-flex flex-grow-1">
         <SidePanel currentTab={currentTab} />
 
@@ -61,9 +48,7 @@ function HomePage() {
       </div>
       <SiteFooter />
 
-      <AboutModal show={shouldAboutShowModal} setShow={shouldShowAboutModal} />
       <ContactModal show={showContactModal} setShow={shouldShowContactModal} />
-      <TermsModal show={showTermsModal} setShow={shouldShowTermsModal} />
     </div>
   );
 }
